+++ cmd/mailchain/internal/rest/handlers/messages.go
@@ -23,8 +23,8 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/gorilla/mux"
-	"github.com/mailchain/mailchain/internal/pkg/encoding"
 	"github.com/mailchain/mailchain/cmd/mailchain/internal/rest/errs"
+	"github.com/mailchain/mailchain/internal/pkg/encoding"
 	"github.com/mailchain/mailchain/internal/pkg/keystore"
 	"github.com/mailchain/mailchain/internal/pkg/keystore/kdf/multi"
 	"github.com/mailchain/mailchain/internal/pkg/mailbox"
