+++ cmd/mailchain/internal/rest/spec/docs.go
