+++ cmd/mailchain/internal/rest/spec/errors.go
