+++ cmd/mailchain/internal/rest/spec/openapi.go
