+++ cmd/mailchain/internal/rest/spec/spec.go
